Transfert de fonction de excel2wisxmlutils vers excel2wisxml
